backend/requirements.txt
fastapi~=0.115.12
python-dotenv~=1.1.0
uvicorn~=0.34.2
pandas~=2.2.3
numpy~=2.0.2
pytz~=2025.2
pydantic~=2.11.4
APScheduler~=3.11.0